prices for events

+++ lib/adminService/EventModel.dart
@@ -43,8 +43,8 @@ class Event {
       singleRooms: data['singleRooms'],
       doubleRooms: data['doubleRooms'],
       description: data['description'],
-      pricePerSingleRoomPerNight: data['pricePerSingleRoomPerNight'] is int ? data['pricePerSingleRoomPerNight'] : int.tryParse(data['pricePerSingleRoomPerNight'].toString()) ?? 0,
-      pricePerDoubleRoomPerNight: data['pricePerDoubleRoomPerNight'] is int ? data['pricePerDoubleRoomPerNight'] : int.tryParse(data['pricePerDoubleRoomPerNight'].toString()) ?? 0,
+      pricePerSingleRoomPerNight: data['pricePerSingleRoomPerNight'] is double ? data['pricePerSingleRoomPerNight'] : double.tryParse(data['pricePerSingleRoomPerNight'].toString()) ?? 0,
+      pricePerDoubleRoomPerNight: data['pricePerDoubleRoomPerNight'] is double ? data['pricePerDoubleRoomPerNight'] : double.tryParse(data['pricePerDoubleRoomPerNight'].toString()) ?? 0,
       imageURLs: List<String>.from(data['imageURLs'] ?? []),
     );
   }
